+++ django_file_form/tus/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
-from .views import TusUpload
+from .views import StartTusUpload, TusUpload
 
 urlpatterns = [
-    path('', TusUpload.as_view(), name='tus_upload'),
+    path('', StartTusUpload.as_view(), name='tus_upload'),
     path('<str:resource_id>', TusUpload.as_view(), name='tus_upload_chunks'),
 ]
